Fixes #3
Should fix https://github.com/rexmtorres/Custom-Build-Scripts/issues/3

As indicated in https://discuss.gradle.org/t/javadoc-exclusion-question/11875/3, it seems that the "exclude" property of Gradle's Javadoc task does not behave in the same way as the "exclude" option of the javadoc executable.  So specifying "**/subpackage/**.java" will not work.

As a work-around to Gradle's Javadoc task issue, we will examine each file in the specified sources and check if the file matches any of the "exclude" patterns.  If it matches, we will filter it out.