Set Ubuntu distribution to "trusty"
To (attempt to) solve the JDK issue.

Apparently, the default JDK being used as of this writing is now JDK 11, which is not compatible with my build.  Setting it to JDK 8 gives another error (Expected feature release number in range of 9 to 14, but got: 8).  This seems to be caused by the default distribution used (Xenial). (source: https://travis-ci.community/t/error-installing-oraclejdk8-expected-feature-release-number-in-range-of-9-to-14-but-got-8/3766/6)